Improve node moving behaviour
When moving selected nodes with ctrl + (up, down, left, right), a more
sophisticated algorithm is now used to determine which other nodes to
'push' along with the selected ones.

Within the interlinked groups which include the selected nodes, the
minimum number of other nodes are 'pushed' along to avoid overlaps. This
is the same as the previous behavior.

However, for interlinked groups that don't include selected nodes, when
they need to be 'pushed' out of the way, the whole group is moved as one.
This prevents the internal movement of the group(s) the user is focusing
on from upsetting the layout of other unrelated groups.